Move txlock.go to unison/txlock.go
